refactoring folder and add new items
