+++ lib/Screens/components/pageitems/home_items.dart
@@ -1,13 +1,13 @@
 import 'package:statusvaccini/models/opendata.dart';
-import 'package:statusvaccini/screens/components/graph_bar_card.dart';
-import 'package:statusvaccini/screens/components/graph_linear_card.dart';
-import 'package:statusvaccini/screens/components/graph_multiple_linear_card.dart';
-import 'package:statusvaccini/screens/components/regioni_card_view.dart';
-import 'package:statusvaccini/screens/components/graph_linear_ultime_consegne.dart';
-import 'package:statusvaccini/screens/components/graph_linear_ultime_sommistrazioni.dart';
+import 'package:statusvaccini/screens/components/graphs/graph_bar_card.dart';
+import 'package:statusvaccini/screens/components/graphs/graph_linear_card.dart';
+import 'package:statusvaccini/screens/components/graphs/graph_multiple_linear_card.dart';
+import 'package:statusvaccini/screens/components/cards/regioni_card_view.dart';
+import 'package:statusvaccini/screens/components/graphs/graph_linear_ultime_consegne.dart';
+import 'package:statusvaccini/screens/components/graphs/graph_linear_ultime_sommistrazioni.dart';
 import 'package:flutter/material.dart';
 
-import 'graph_pie_card.dart';
+import '../graphs/graph_pie_card.dart';
 
 class HomeItems {
   Widget card;
