+++ lib/Screens/views/home_view.dart
@@ -1,4 +1,4 @@
-import 'package:statusvaccini/screens/components/home_items.dart';
+import 'package:statusvaccini/screens/components/pageitems/home_items.dart';
 import 'package:statusvaccini/constants/constant.dart';
 import 'package:flutter/material.dart';
 import 'package:statusvaccini/screens/components/body_components.dart';
